feat!: withPlugins utility
