+++ .changeset/mean-windows-tan.md
@@ -1,0 +1,5 @@
+---
+"astro-integration-kit": minor
+---
+
+Plugins are not longer passed to `defineIntegration`. Instead, there's a new `withPlugins` utility. **This is a breaking change**, check out the [migration guide](https://astro-integration-kit.netlify.app/getting-started/breaking-changes/#090)
