+++ docs/src/content/docs/core/with-plugins.mdx
@@ -1,0 +1,24 @@
+---
+title: withPlugins
+description: Allows to extend hooks with custom parameters. Advanced.
+---
+
+`withPlugins` allows to extend hooks with custom parameters. Advanced.
+
+```ts title="my-integration/index.ts" "withPlugins"
+import { defineIntegration, withPlugins } from "astro-integration-kit";
+import { hasVitePluginPlugin  } from "astro-integration-kit/plugins";
+
+export default defineIntegration({
+    name: "my-integration",
+    setup({ name }) {
+        return withPlugins({
+            name,
+            plugins: [hasVitePluginPlugin],
+            hooks: {
+                "astro:config:setup": ({ hasVitePluginPlugin }) => {}
+            }
+        })
+    }
+})
+```
