+++ docs/src/content/docs/getting-started/breaking-changes.mdx
@@ -7,6 +7,39 @@ import { Tabs, TabItem } from '@astrojs/starlight/components';
 
 Features get added and removed, and breaking changes are introduced! This documents how to migrate.
 
+## `0.9.0`
+
+---
+
+### New `withPlugins` utility, `plugins` removed from `defineIntegration`
+
+Plugins are not longer passed to `defineIntegration`. Instead, there's a new `withPlugins` utility.
+
+```ts title="my-integration/index.ts" del={9,11-13} ins={3,14-20}
+import {
+    defineIntegration,
+    withPlugins
+} from "astro-integration-kit";
+import { hasVitePluginPlugin } from "astro-integration-kit/plugins";
+
+export default defineIntegration({
+    name: "my-integration",
+    plugins: [hasVitePluginPlugin],
+    setup({ name }) {
+        return {
+            "astro:config:setup": ({ hasVitePluginPlugin }) => {}
+        }
+        return withPlugins({
+            name,
+            plugins: [hasVitePluginPlugin],
+            hooks: {
+                "astro:config:setup": ({ hasVitePluginPlugin }) => {}
+            }
+        })
+    }
+})
+```
+
 ## `0.8.0`
 
 ---
