+++ package/src/core/index.ts
@@ -2,5 +2,6 @@ export { createResolver } from "./create-resolver.js";
 export { defineIntegration } from "./define-integration.js";
 export { definePlugin } from "./define-plugin.js";
 export { defineUtility } from "./define-utility.js";
+export { withPlugins } from "../core/with-plugins.js";
 export * from "./types.js";
 export * from "../utilities/index.js";
